feat: Phase 10 bundle — core package, git-cliff, contracts.tools, cleanup
- New @karmaniverous/jeeves-meta-core package exporting metaConfigSchema + MetaConfig
- Import metaConfigSchema from core in service
- Migrated changelog from auto-changelog to git-cliff in service and openclaw
- Added contracts.tools to openclaw.plugin.json
- Bumped @karmaniverous/jeeves to ^0.5.10
- Applied safe dependency updates via ncu
- Added vitest.config.ts to core and openclaw for proper test resolution
