+++ packages/core/.env.local.template
@@ -1,0 +1,13 @@
+# This file is a template for the .env.local. Add your Personal Access Token
+# below to streamline the release process from your desktop! Just follow these
+# steps:
+#
+# 1. Make a copy of this file and rename it to .env.local.
+#
+# 2. Create a Personal Access Token at https://github.com/settings/tokens and
+#    enter it below.
+#
+# .env.local is gitignored and will NOT be persisted to your repo, but don't
+# put any secrets in this template!
+
+GITHUB_TOKEN=
