+++ packages/core/src/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Shared core types and schemas for jeeves-meta packages.
+ *
+ * @packageDocumentation
+ */
+
+export { type MetaConfig, metaConfigSchema } from './metaConfig.js';
