+++ packages/openclaw/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "types": ["node"]
   },
-  "include": ["src/**/*", "scripts/**/*", "rollup.config.ts"],
+  "include": ["src/**/*", "scripts/**/*", "rollup.config.ts", "vitest.config.ts"],
   "exclude": ["dist", "node_modules"]
 }
