+++ packages/service/package.json
@@ -41,33 +41,33 @@
     "node": ">=22"
   },
   "dependencies": {
-    "@karmaniverous/jeeves": "^0.5.9",
+    "@karmaniverous/jeeves": "^0.5.10",
+    "@karmaniverous/jeeves-meta-core": "^0.1.0",
     "commander": "^14",
     "croner": "^10",
     "fastify": "^5.8",
     "handlebars": "^4.7.9",
     "package-directory": "^8.2.0",
     "pino": "^10",
-    "zod": "^4.3"
+    "zod": "^4.4"
   },
   "devDependencies": {
     "@commander-js/extra-typings": "^14.0.0",
-    "@dotenvx/dotenvx": "^1.59.1",
+    "@dotenvx/dotenvx": "^1.65.0",
     "@rollup/plugin-commonjs": "^29.0.2",
     "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^16.0.3",
     "@rollup/plugin-typescript": "^12.3.0",
-    "@types/node": "^25.5.2",
-    "@vitest/coverage-v8": "^4.1.2",
-    "auto-changelog": "^2.5.0",
+    "@types/node": "^25.7.0",
+    "@vitest/coverage-v8": "^4.1.6",
     "cross-env": "^10.1.0",
     "knip": "^5.87.0",
     "release-it": "^19.2.4",
-    "rollup": "^4.60.1",
+    "rollup": "^4.60.3",
     "rollup-plugin-copy": "^3.5.0",
     "rollup-plugin-dts": "^6.4.1",
     "tslib": "^2.8.1",
-    "vitest": "^4.1.2"
+    "vitest": "^4.1.6"
   },
   "scripts": {
     "build": "rimraf dist && cross-env NO_COLOR=1 rollup --config rollup.config.ts --configPlugin @rollup/plugin-typescript",
@@ -76,7 +76,7 @@
     "lint:fix": "eslint --fix .",
     "test": "vitest run",
     "typecheck": "tsc",
-    "changelog": "auto-changelog",
+    "changelog": "git-cliff -o CHANGELOG.md",
     "release": "dotenvx run -f .env.local -- release-it",
     "release:pre": "dotenvx run -f .env.local -- release-it --no-git.requireBranch --github.prerelease --preRelease"
   },
@@ -85,7 +85,7 @@
   },
   "release-it": {
     "git": {
-      "changelog": "npx auto-changelog --unreleased-only --stdout --template https://raw.githubusercontent.com/release-it/release-it/main/templates/changelog-compact.hbs",
+      "changelog": "npx git-cliff --unreleased --strip header",
       "commitMessage": "chore: release @karmaniverous/jeeves-meta v${version}",
       "tagName": "service/${version}",
       "requireBranch": "main"
@@ -105,20 +105,13 @@
         "git push -u origin release/service/${version}",
         "git switch ${branchName}"
       ],
-      "after:bump": [
-        "npx auto-changelog -p",
-        "git add CHANGELOG.md"
+      "before:npm:release": [
+        "npx git-cliff -o CHANGELOG.md",
+        "git add -A"
       ]
     },
     "npm": {
       "publish": true
     }
-  },
-  "auto-changelog": {
-    "output": "CHANGELOG.md",
-    "tagPrefix": "service/",
-    "unreleased": true,
-    "commitLimit": false,
-    "hideCredit": true
   }
 }
